// frontend/frontend/src/app/models/Class.ts
export class Class {
    id: number = 0;
    student: string = "";
    tutor: string = "";
    subject: string = "";
    description: string = "";
    datetime: Date = new Date();
    doubletime: boolean = false;
    approved: number = 0;
    explanation: string = "";
    reviewForTutor: number = 0;
    commentForTutor: string = "";
    reviewForStudent: number = 0;
    commentForStudent: string = "";
}